A dto used for response of user registering API. Spring's conversion service used to convert types. Some refactorings done because of some packaging.
