--- src/main/java/com/softwaretalks/jangul/controllers/UserController.java
@@ -1,25 +1,0 @@
-package com.softwaretalks.jangul.controllers;
-
-import com.softwaretalks.jangul.models.User;
-import com.softwaretalks.jangul.repositories.UserRepository;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public UserController(UserRepository userRepository, PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    @PostMapping("/users")
-    public User postUser(@RequestBody User user) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        return userRepository.save(user);
-    }
-}
