+++ src/main/java/com/softwaretalks/jangul/web/dto/user/UserDto.java
@@ -1,0 +1,16 @@
+package com.softwaretalks.jangul.web.dto.user;
+
+
+import lombok.Getter;
+
+/**
+ * Represents attributes to be displayed in registering a new user.
+ *
+ * @author sajad
+ */
+
+@Getter
+public class UserDto {
+
+    private String username;
+}
